item save

+++ backend/database/migrations/2023_06_17_141741_create_item_prices_table.php
@@ -14,7 +14,7 @@ return new class extends Migration {
             $table->id();
             $table->timestamps();
             $table->bigInteger('item_id')->nullable();
-            $table->float('price')->nullable();
+            $table->double('price')->nullable();
         });
     }
